refactor: move LocalRunnerScanner to RunnerBarCore/Runner (#924)
- Delete Sources/RunnerBar/Runner/LocalRunnerScanner.swift
- Add Sources/RunnerBarCore/Runner/LocalRunnerScanner.swift with public API
- Remove import RunnerBarCore (can't import yourself)
- ProcessRunner.run() now resolved within Core directly
- No import changes needed in callers (covered by @_exported in Exports.swift)
